meson: pass library install paths to g-ir-scanner
We are patching g-ir-scanner to produce absolute paths in the GIR
files. When an application uses an internal library placed in a
non-standard path (e.g. $out/lib/gnome-shell), the scanner needs
to be informed. For autotools-based apps, the full path was
obtained from libtool-wrapped files; with Meson, this is no longer
possible – we need to pass the path information some way.

This commit channels the --fallback-library-path option added
to g-ir-scanner in aforementioned patch.